requirements.txt
mss
numpy
opencv-python
Pillow
pydirectinput-rgx
pywinctl
rich
tessdata.fast-eng
tesserocr
tqdm
black
